Move utility functions in to an internal directory
Debates about "util" packages aside, this is a mechanical cleanup to remove external
surface area.
