+++ pkg/backends/statsdaemon/statsdaemon.go
@@ -15,9 +15,9 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/atlassian/gostatsd"
+	"github.com/atlassian/gostatsd/internal/util"
 	"github.com/atlassian/gostatsd/pkg/backends/sender"
 	"github.com/atlassian/gostatsd/pkg/transport"
-	"github.com/atlassian/gostatsd/pkg/util"
 )
 
 const (
