+++ pkg/cachedinstances/k8s/k8s.go
@@ -19,7 +19,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/atlassian/gostatsd"
-	"github.com/atlassian/gostatsd/pkg/util"
+	"github.com/atlassian/gostatsd/internal/util"
 )
 
 var (
